requirementes.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
elastic-transport==8.4.0
elasticsearch==8.6.1
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
